+++ base/set.jl
@@ -1,8 +1,8 @@
 # This file is a part of Julia. License is MIT: https://julialang.org/license
 
-eltype(::Type{AbstractSet{T}}) where {T} = T
+eltype(::Type{<:AbstractSet{T}}) where {T} = @isdefined(T) ? T : Any
 
-mutable struct Set{T} <: AbstractSet{T}
+struct Set{T} <: AbstractSet{T}
     dict::Dict{T,Nothing}
 
     Set{T}() where {T} = new(Dict{T,Nothing}())