+++ stdlib/Base64/src/Base64.jl
@@ -8,7 +8,8 @@ export
     Base64EncodePipe,
     base64encode,
     Base64DecodePipe,
-    base64decode
+    base64decode,
+    stringmime
 
 # Base64EncodePipe is a pipe-like IO object, which converts into base64 data
 # sent to a stream. (You must close the pipe to complete the encode, separate
@@ -22,5 +23,22 @@ export
 include("buffer.jl")
 include("encode.jl")
 include("decode.jl")
+
+"""
+    stringmime(mime, x; context=nothing)
+
+Returns an `AbstractString` containing the representation of `x` in the
+requested `mime` type. This is similar to [`repr(mime, x)`](@ref) except
+that binary data is base64-encoded as an ASCII string.
+
+The optional keyword argument `context` can be set to `:key=>value` pair
+or an `IO` or [`IOContext`](@ref) object whose attributes are used for the I/O
+stream passed to `show`.
+"""
+stringmime(m::MIME, x; context=nothing) = istextmime(m) ? Base.Multimedia._textrepr(m, x, context) : _binstringmime(m, x, context)
+stringmime(m::AbstractString, x; context=nothing) = stringmime(MIME(m), x; context=context)
+
+_binstringmime(m::MIME, x, context) = Base64.base64encode(show, m, x; context=IOContext)
+_binstringmime(m::MIME, x::Vector{UInt8}, context) = Base64.base64encode(write, x; context=context)
 
 end