+++ stdlib/LinearAlgebra/src/dense.jl
@@ -72,7 +72,7 @@ end
 """
     isposdef!(A) -> Bool
 
-Test whether a matrix is positive definite by trying to perform a
+Test whether a matrix is positive definite (and Hermitian) by trying to perform a
 Cholesky factorization of `A`, overwriting `A` in the process.
 See also [`isposdef`](@ref).
 
@@ -94,7 +94,7 @@ isposdef!(A::AbstractMatrix) = ishermitian(A) && isposdef(cholfact!(Hermitian(A)
 """
     isposdef(A) -> Bool
 
-Test whether a matrix is positive definite by trying to perform a
+Test whether a matrix is positive definite (and Hermitian) by trying to perform a
 Cholesky factorization of `A`.
 See also [`isposdef!`](@ref)
 
@@ -249,7 +249,7 @@ function diagind(m::Integer, n::Integer, k::Integer=0)
         throw(ArgumentError(string("requested diagonal, $k, must be at least $(-m) and ",
             "at most $n in an $m-by-$n matrix")))
     end
-    k <= 0 ? range(1-k, m+1, min(m+k, n)) : range(k*m+1, m+1, min(m, n-k))
+    k <= 0 ? range(1-k, step=m+1, length=min(m+k, n)) : range(k*m+1, step=m+1, length=min(m, n-k))
 end
 
 """