+++ stdlib/SHA/src/SHA.jl
@@ -1,4 +1,4 @@
-__precompile__()
+__precompile__(true)
 
 module SHA